fix: close 6 correctness gaps found tracing architecture.md step-by-step
Traced docs/diagrams/architecture.md's layer boundaries (api/config, data,
core physics, engine, service seam, front ends, results/viz/GUI) against
the real code -- 5 forks per subgraph plus Agy CLI review (Codex was
quota-limited) on the core physics kernels/models -- and fixed all
adversarially-confirmed defects:

- homodyne_model.py::_extract_config and physics_factors.py::
  create_physics_factors_from_config_dict both hard-required a nested
  analyzer_parameters.temporal.{dt,start_frame,end_frame} shape, but the
  canonical shipped YAML templates (xpcsjax/config/templates/*.yaml) put
  these FLAT under analyzer_parameters -- constructing HomodyneModel
  directly from a real template-derived config raised KeyError. Now
  accepts the flat form (canonical) with the nested form as a fallback,
  matching HeterodyneModel.from_config's existing pattern.
- models.py::make_model dispatched the homodyne path via the strict
  AnalysisMode(mode_lower) constructor instead of AnalysisMode.parse(),
  rejecting documented synonyms like "laminar" that .parse() accepts.
- heterodyne_model_stateful.py::create_residual_function built
  varying_idx_jax via jnp.array(...) with no explicit dtype; an empty
  varying_indices list (every physical parameter fixed) defaults to
  float64, and .at[float_index].set(...) raises TypeError.
- viz/nlsq_plots.py: plot_nlsq_fit/plot_residual_map/plot_simulated_data
  built a matplotlib Figure with no try/finally before rendering; any
  mid-render exception (bad data reaching imshow/colorbar/percentile
  calcs) leaked the Figure, accumulating across a batch per-angle loop
  where every production caller catches-and-continues on error.
  datashader_backend.py already closes correctly -- these three were the
  outliers.
- docs/diagrams/architecture.md: corrected a stale DQC-->MEM edge
  (DataQualityController does not hand off to AdvancedMemoryManager;
  both are wired independently from load_xpcs_data).

Investigated and NOT changed (false positives / out of scope):
- physics_nlsq.py contrast/offset array-broadcast concern: the type hint
  allows an array, but no caller ever passes one (always scalar per-angle).
- heterodyne_models.py TwoComponentModel.compute_g1_{reference,sample}:
  flagged pointwise formula is a documented non-production visualization
  helper (docstring says so explicitly), not the production integral path.
- physics.py get_default_parameters()'s gamma_dot_t_offset default
  (SHEAR_OFFSET_MIN) yields a negative net initial shear-rate guess --
  real but a physics-default judgment call, left for a maintainer decision
  rather than a unilateral change.
- CM-->NCFG edge in the old diagram was already stale (ConfigManager
  never constructs NLSQConfig); not otherwise a code defect.

Verified: ruff clean, tests/core/ 73/73, tests/viz/ 111/111, full smoke
suite ran clean to ~95-98% across two runs before external interruption
(session restarts killed the background process, not a test failure).

+++ xpcsjax/core/heterodyne_model_stateful.py
@@ -453,7 +453,11 @@ class HeterodyneModel:
 
         contrast_val, offset_val = self.scaling.get_for_angle(angle_idx)
 
-        varying_idx_jax = jnp.array(self.param_manager.varying_indices)
+        # Explicit int dtype: jnp.array([]) on an empty varying_indices list
+        # (e.g. every physical parameter fixed) defaults to float64, and a
+        # float-dtype .at[...].set() indexer raises TypeError ("Indexer must
+        # have integer or boolean type").
+        varying_idx_jax = jnp.array(self.param_manager.varying_indices, dtype=jnp.int32)
         fixed_values_jax = jnp.array(self.param_manager.get_full_values())
 
         @jax.jit
